+++ ontime-back/src/main/java/devkor/ontime_back/dto/OAuthGoogleRequestDto.java
@@ -5,4 +5,5 @@ import lombok.Getter;
 @Getter
 public class OAuthGoogleRequestDto {
     private String accessToken;
+    private String refreshToken;
 }
